YUK-11772 Remove net.sf.jsonOLD
Removed from user preferences, SupportController, and LogExplorerController

+++ yukon-client/yukon-web/src/com/cannontech/web/user/validator/ChangePasswordValidatorFactory.java
@@ -74,7 +74,7 @@ public class ChangePasswordValidatorFactory {
     
         public void validate(ChangePassword input, Errors errors) {
     
-            final LiteYukonUser user = yukonUserDao.getLiteYukonUser(input.getUserId().intValue());
+            final LiteYukonUser user = yukonUserDao.getLiteYukonUser(input.getUserId());
     
             UserAuthenticationInfo userAuthenticationInfo = yukonUserDao.getUserAuthenticationInfo(user.getUserID());
             AuthenticationCategory authenticationCategory = userAuthenticationInfo.getAuthenticationCategory();
